requirements.txt
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.0
cryptography==3.4.6
flake8==4.0.1
idna==2.10
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.6
python-dotenv==0.20.0
requests==2.25.1
service-identity==18.1.0
six==1.15.0
urllib3==1.26.3
